# docs/swagger.yaml
basePath: /
definitions:
  backends.Backend:
    properties:
      address:
        example: 127.0.0.1:80
        type: string
      site_id:
        $ref: '#/definitions/sites.Site'
    type: object
...
  sites.Site:
    properties:
      host:
        example: site.com
        type: string
      name:
        example: site
        type: string
    type: object
host: localhost:80
info:
  contact: {}
  description: How to work CRUD server in reverseProxy
  title: CRUD server in reverseProxy
  version: 1.0.0
paths:
  /backends:
    post:
      consumes:
      - application/json
      description: Create backends
      parameters:
      - description: backend info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/backends.Backend'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: string
      summary: Create new backends
      tags:
      - Backends
  /backends/{id}:
    delete:
      consumes:
      - application/json
      description: delete backends
      parameters:
      - description: backends ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backends.Backend'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Delete backends based on given id
      tags:
      - Backends
    get:
      consumes:
      - application/json
      description: get backends
      parameters:
      - description: backends ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backends.Backend'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get backends based on given id
      tags:
      - Backends
    put:
      consumes:
      - application/json
      description: update backends
      parameters:
      - description: backends ID
        in: path
        name: id
        required: true
        type: integer
      - description: backends info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/backends.Backend'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backends.Backend'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Update backends based on given id
      tags:
      - Backends
...
schemes:
- http
swagger: "2.0"